+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - orderBy Step can generate ASC, DESC, NULLS FIRST, NULLS LAST
 - 8 functions added to Column class to give OrderByItemInfo class
 - Column class knows its table
+- Add **Distinct** clause
 
 ## [0.5.1]  2022-01-29
 ### fixed